refactor: move categories fetcher
