+++ app/components/Header/useSearch.ts
@@ -1,8 +1,12 @@
-import { Category, Link } from "@/app/[[...category]]/get-data";
+import { Category, Link } from "@/app/utils/categories";
+
 import { useMemo, useState } from "react";
 import lunr from "lunr";
+// @ts-ignore
 import withStemmerSupport from "lunr-languages/lunr.stemmer.support";
+// @ts-ignore
 import withHebrewSupport from "lunr-languages/lunr.he";
+// @ts-ignore
 import withMulti from "lunr-languages/lunr.multi";
 import mixpanel from "mixpanel-browser";
