+++ app/utils/categories.ts
